Add an effect
The dependency array at the end of this is interesting. You add all the
external variables being set to it, and if any of those change, the
effect will run. This is meant to be there so it doesn't rerun anymore
than you want it to.

From what I can tell there are three "modes" you can set this to.

First is just no including it at all, which causes the effect to run
after every render.

Second is by adding the dependencies that are used inside the effect.
Linters will tell you what these are if you specify at least one
argument inside the array.

Third is passing an empty array. This causes the effect to run after the
component mounts for the first time, but not on any of the subsequent
rerenders.